+++ packages/component/src/decorator/private/BorderLoader.tsx
@@ -3,15 +3,20 @@ import { useStyles } from 'botframework-webchat-styles/react';
 
 import styles from './BorderLoader.module.css';
 
-function BorderLoader({ children }: Readonly<{ children?: ReactNode | undefined }>) {
+function BorderLoader({
+  children,
+  showLoader = true
+}: Readonly<{ children?: ReactNode | undefined; showLoader?: boolean }>) {
   const classNames = useStyles(styles);
 
   return (
     <div className={classNames['border-loader']}>
       {children}
-      <div className={classNames['border-loader__track']}>
-        <div className={classNames['border-loader__loader']} />
-      </div>
+      {showLoader && (
+        <div className={classNames['border-loader__track']}>
+          <div className={classNames['border-loader__loader']} />
+        </div>
+      )}
     </div>
   );
 }